Refactor test boilerplate to support multiple solr versions.
